feat(test): added pwd tests + some refactor
