+++ src/test/java/org/itmo/modules/LocalStorageTests.java
@@ -1,6 +1,5 @@
-package modules;
+package org.itmo.modules;
 
-import org.itmo.modules.LocalStorage;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
