+++ src/test/java/org/itmo/modules/ParserTests.java
@@ -1,6 +1,5 @@
-package modules;
+package org.itmo.modules;
 
-import org.itmo.modules.Parser;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
